Add empty cata module; move tree to Eval.Types
